+++ buession-logging-support-rabbitmq/src/main/java/com/buession/logging/rabbitmq/spring/ConnectionFactoryBean.java
@@ -50,7 +50,9 @@ public class ConnectionFactoryBean extends ConnectionFactory
 
 	@Override
 	public void afterPropertiesSet() throws Exception {
-		connectionFactory = createConnectionFactory();
+		if(connectionFactory == null){
+			connectionFactory = createConnectionFactory();
+		}
 	}
 
 	@Override